+++ cedar-policy-core/src/ast/policy_set.rs
@@ -25,6 +25,12 @@ use std::collections::{hash_map::Entry, HashMap, HashSet};
 use std::{borrow::Borrow, sync::Arc};
 use thiserror::Error;
 
+#[cfg(feature = "protobufs")]
+use crate::ast::proto;
+
+#[cfg(feature = "protobufs")]
+use super::TemplateBody;
+
 /// Represents a set of `Policy`s
 #[derive(Debug, Default, Clone, PartialEq, Eq, Serialize, Deserialize)]
 #[serde(try_from = "LiteralPolicySet")]
@@ -106,6 +112,94 @@ impl From<PolicySet> for LiteralPolicySet {
             .map(|(id, p)| (id, p.into()))
             .collect();
         Self { templates, links }
+    }
+}
+
+#[cfg(feature = "protobufs")]
+impl From<&proto::LiteralPolicySet> for LiteralPolicySet {
+    fn from(v: &proto::LiteralPolicySet) -> Self {
+        let templates: HashMap<PolicyID, Template> = v
+            .templates
+            .iter()
+            .map(|(key, value)| {
+                (
+                    PolicyID::from_string(key),
+                    Template::from(TemplateBody::from(value)),
+                )
+            })
+            .collect();
+
+        let links: HashMap<PolicyID, LiteralPolicy> = v
+            .links
+            .iter()
+            .map(|(key, value)| (PolicyID::from_string(key), LiteralPolicy::from(value)))
+            .collect();
+
+        Self {
+            templates: templates,
+            links: links,
+        }
+    }
+}
+
+#[cfg(feature = "protobufs")]
+impl From<&LiteralPolicySet> for proto::LiteralPolicySet {
+    fn from(v: &LiteralPolicySet) -> Self {
+        let mut templates: HashMap<String, proto::TemplateBody> =
+            HashMap::with_capacity(v.templates.len());
+        for (key, value) in &v.templates {
+            templates.insert(String::from(key.as_ref()), proto::TemplateBody::from(value));
+        }
+
+        let mut links: HashMap<String, proto::LiteralPolicy> =
+            HashMap::with_capacity(v.links.len());
+        for (key, value) in &v.links {
+            links.insert(
+                String::from(key.as_ref()),
+                proto::LiteralPolicy::from(value),
+            );
+        }
+
+        Self {
+            templates: templates,
+            links: links,
+        }
+    }
+}
+
+#[cfg(feature = "protobufs")]
+impl From<&PolicySet> for proto::LiteralPolicySet {
+    fn from(v: &PolicySet) -> Self {
+        let mut templates: HashMap<String, proto::TemplateBody> =
+            HashMap::with_capacity(v.templates.len());
+        for (key, value) in &v.templates {
+            templates.insert(
+                String::from(key.as_ref()),
+                proto::TemplateBody::from(value.as_ref()),
+            );
+        }
+
+        let mut links: HashMap<String, proto::LiteralPolicy> =
+            HashMap::with_capacity(v.links.len());
+        for (key, value) in &v.links {
+            links.insert(
+                String::from(key.as_ref()),
+                proto::LiteralPolicy::from(value),
+            );
+        }
+
+        Self {
+            templates: templates,
+            links: links,
+        }
+    }
+}
+
+#[cfg(feature = "protobufs")]
+impl TryFrom<&proto::LiteralPolicySet> for PolicySet {
+    type Error = ReificationError;
+    fn try_from(pset: &proto::LiteralPolicySet) -> Result<Self, Self::Error> {
+        PolicySet::try_from(LiteralPolicySet::from(pset))
     }
 }
 
@@ -511,7 +605,17 @@ mod test {
         },
         parser,
     };
+
     use std::collections::HashMap;
+
+    #[cfg(feature = "protobufs")]
+    use crate::{
+        ast::{Annotation, AnyId, EntityType, Name, PrincipalOrResourceConstraint},
+        from_normalized_str::FromNormalizedStr,
+    };
+
+    #[cfg(feature = "protobufs")]
+    use std::str::FromStr;
 
     #[test]
     fn link_conflicts() {
@@ -910,5 +1014,114 @@ mod test {
         );
         assert!(pset.get(&tid1).is_none());
         assert_eq!(pset.all_templates().count(), 4);
+    }
+
+    #[cfg(feature = "protobufs")]
+    #[test]
+    fn protobuf_roundtrip() {
+        let annotation1: Annotation = Annotation {
+            val: "".into(),
+            loc: None,
+        };
+        let pc: PrincipalConstraint =
+            PrincipalConstraint::is_eq(EntityUID::with_eid("friend").into());
+        let ac: ActionConstraint = ActionConstraint::Eq(EntityUID::with_eid("read").into());
+        let rc: ResourceConstraint = ResourceConstraint {
+            constraint: PrincipalOrResourceConstraint::is_entity_type(
+                EntityType::from(Name::from_normalized_str("photo").unwrap()).into(),
+            ),
+        };
+
+        let tb: TemplateBody = TemplateBody::new(
+            PolicyID::from_string("template"),
+            None,
+            Annotations::from_iter(vec![(AnyId::from_str("read").expect(""), annotation1)]),
+            Effect::Permit,
+            pc,
+            ac,
+            rc,
+            Expr::val(true),
+        );
+
+        let policy: Policy = Policy::from_when_clause(
+            Effect::Permit,
+            Expr::val(true),
+            PolicyID::from_string("alice"),
+            None,
+        );
+
+        let mut ps: PolicySet = PolicySet::new();
+        ps.add_template(Template::from(tb))
+            .expect("Failed to add template to policy set.");
+        ps.add(policy).expect("Failed to add policy to policy set.");
+        let lps: LiteralPolicySet = LiteralPolicySet::from(ps);
+        let lps_roundtrip: LiteralPolicySet =
+            LiteralPolicySet::from(&proto::LiteralPolicySet::from(&lps));
+
+        // Can't compare LiteralPolicySets directly, so we compare their fields
+        assert_eq!(lps.templates, lps_roundtrip.templates);
+        assert_eq!(lps.links, lps_roundtrip.links);
+    }
+
+    #[cfg(feature = "protobufs")]
+    #[test]
+    fn protobuf_roundtrip_forbids() {
+        let annotation1: Annotation = Annotation {
+            val: "".into(),
+            loc: None,
+        };
+        let pc: PrincipalConstraint =
+            PrincipalConstraint::is_eq(EntityUID::with_eid("friend").into());
+        let ac: ActionConstraint = ActionConstraint::Eq(EntityUID::with_eid("read").into());
+        let rc: ResourceConstraint = ResourceConstraint {
+            constraint: PrincipalOrResourceConstraint::is_entity_type(
+                EntityType::from(Name::from_normalized_str("photo").unwrap()).into(),
+            ),
+        };
+
+        let tb: TemplateBody = TemplateBody::new(
+            PolicyID::from_string("template"),
+            None,
+            Annotations::from_iter(vec![(AnyId::from_str("read").expect(""), annotation1)]),
+            Effect::Forbid,
+            pc,
+            ac,
+            rc,
+            Expr::val(true),
+        );
+
+        let policy: Policy = Policy::from_when_clause(
+            Effect::Permit,
+            Expr::val(true),
+            PolicyID::from_string("alice"),
+            None,
+        );
+
+        let mut ps: PolicySet = PolicySet::new();
+        ps.add_template(Template::from(tb))
+            .expect("Failed to add template to policy set.");
+        ps.add(policy.clone())
+            .expect("Failed to add policy to policy set.");
+        let lps: LiteralPolicySet = LiteralPolicySet::from(ps.clone());
+        let lps_roundtrip: LiteralPolicySet =
+            LiteralPolicySet::from(&proto::LiteralPolicySet::from(&lps));
+
+        // Can't compare LiteralPolicySets directly, so we compare their fields
+        assert_eq!(lps.templates, lps_roundtrip.templates);
+        assert_eq!(lps.links, lps_roundtrip.links);
+
+        ps.remove_static(policy.id()).unwrap();
+        let policy: Policy = Policy::from_when_clause(
+            Effect::Forbid,
+            Expr::val(true),
+            PolicyID::from_string("alice"),
+            None,
+        );
+        ps.add(policy).expect("Failed to add policy to policy set.");
+        let lps: LiteralPolicySet = LiteralPolicySet::from(ps);
+        // Static policies have templates, so not equal
+        assert_ne!(lps.templates, lps_roundtrip.templates);
+        // The static policies are identical except for the template, so links are equal
+        assert_eq!(lps.links, lps_roundtrip.links);
     }
 }